+++ vite.config.js
@@ -4,6 +4,7 @@ import vue from '@vitejs/plugin-vue'
 
 export default defineConfig({
   plugins: [vue()],
+  base: '/kids-timer/',
   server: {
     port: 3000
   }